fix(review): remove completion tool import cycle
